[worker-api] move definition of `isPubKeyPinPair` to common
